Refactor GoGame into multiple classes
Add back newline

Fix broadcasting
