+++ app/models/go.rb
@@ -1,0 +1,3 @@
+# namespace for the related Go classes
+module Go
+end
